Refactor configs package
- Rename Config to ConfigParams
- Move parsing-related common functions into  parsing_helpers.go
- Move Ingress-related functions into ingress.go
- Rename templates package to version1
- Rename nginx.go into config.go and move to version1 package
- Move Configmaps-related functions configmaps.go
- Move annotations-related functions to annotations.go
- Move secrets.go from configs to k8s
- Introduce StaticConfigParams
- Move template_executor.go to version1
