+++ internal/configs/annotations_test.go
@@ -1,0 +1,173 @@
+package configs
+
+import (
+	"reflect"
+	"sort"
+	"testing"
+)
+
+func TestParseRewrites(t *testing.T) {
+	serviceName := "coffee-svc"
+	serviceNamePart := "serviceName=" + serviceName
+	rewritePath := "/beans/"
+	rewritePathPart := "rewrite=" + rewritePath
+	rewriteService := serviceNamePart + " " + rewritePathPart
+
+	serviceNameActual, rewritePathActual, err := parseRewrites(rewriteService)
+	if serviceName != serviceNameActual || rewritePath != rewritePathActual || err != nil {
+		t.Errorf("parseRewrites(%s) should return %q, %q, nil; got %q, %q, %v", rewriteService, serviceName, rewritePath, serviceNameActual, rewritePathActual, err)
+	}
+}
+
+func TestParseRewritesWithLeadingAndTrailingWhitespace(t *testing.T) {
+	serviceName := "coffee-svc"
+	serviceNamePart := "serviceName=" + serviceName
+	rewritePath := "/beans/"
+	rewritePathPart := "rewrite=" + rewritePath
+	rewriteService := "\t\n " + serviceNamePart + " " + rewritePathPart + " \t\n"
+
+	serviceNameActual, rewritePathActual, err := parseRewrites(rewriteService)
+	if serviceName != serviceNameActual || rewritePath != rewritePathActual || err != nil {
+		t.Errorf("parseRewrites(%s) should return %q, %q, nil; got %q, %q, %v", rewriteService, serviceName, rewritePath, serviceNameActual, rewritePathActual, err)
+	}
+}
+
+func TestParseRewritesInvalidFormat(t *testing.T) {
+	rewriteService := "serviceNamecoffee-svc rewrite=/"
+
+	_, _, err := parseRewrites(rewriteService)
+	if err == nil {
+		t.Errorf("parseRewrites(%s) should return error, got nil", rewriteService)
+	}
+}
+
+func TestParseStickyService(t *testing.T) {
+	serviceName := "coffee-svc"
+	serviceNamePart := "serviceName=" + serviceName
+	stickyCookie := "srv_id expires=1h domain=.example.com path=/"
+	stickyService := serviceNamePart + " " + stickyCookie
+
+	serviceNameActual, stickyCookieActual, err := parseStickyService(stickyService)
+	if serviceName != serviceNameActual || stickyCookie != stickyCookieActual || err != nil {
+		t.Errorf("parseStickyService(%s) should return %q, %q, nil; got %q, %q, %v", stickyService, serviceName, stickyCookie, serviceNameActual, stickyCookieActual, err)
+	}
+}
+
+func TestParseStickyServiceInvalidFormat(t *testing.T) {
+	stickyService := "serviceNamecoffee-svc srv_id expires=1h domain=.example.com path=/"
+
+	_, _, err := parseStickyService(stickyService)
+	if err == nil {
+		t.Errorf("parseStickyService(%s) should return error, got nil", stickyService)
+	}
+}
+
+func TestFilterMasterAnnotations(t *testing.T) {
+	masterAnnotations := map[string]string{
+		"nginx.org/rewrites":                "serviceName=service1 rewrite=rewrite1",
+		"nginx.org/ssl-services":            "service1",
+		"nginx.org/hsts":                    "True",
+		"nginx.org/hsts-max-age":            "2700000",
+		"nginx.org/hsts-include-subdomains": "True",
+	}
+	removedAnnotations := filterMasterAnnotations(masterAnnotations)
+
+	expectedfilteredMasterAnnotations := map[string]string{
+		"nginx.org/hsts":                    "True",
+		"nginx.org/hsts-max-age":            "2700000",
+		"nginx.org/hsts-include-subdomains": "True",
+	}
+	expectedRemovedAnnotations := []string{
+		"nginx.org/rewrites",
+		"nginx.org/ssl-services",
+	}
+
+	sort.Strings(removedAnnotations)
+	sort.Strings(expectedRemovedAnnotations)
+
+	if !reflect.DeepEqual(expectedfilteredMasterAnnotations, masterAnnotations) {
+		t.Errorf("filterMasterAnnotations returned %v, but expected %v", masterAnnotations, expectedfilteredMasterAnnotations)
+	}
+	if !reflect.DeepEqual(expectedRemovedAnnotations, removedAnnotations) {
+		t.Errorf("filterMasterAnnotations returned %v, but expected %v", removedAnnotations, expectedRemovedAnnotations)
+	}
+}
+
+func TestFilterMinionAnnotations(t *testing.T) {
+	minionAnnotations := map[string]string{
+		"nginx.org/rewrites":                "serviceName=service1 rewrite=rewrite1",
+		"nginx.org/ssl-services":            "service1",
+		"nginx.org/hsts":                    "True",
+		"nginx.org/hsts-max-age":            "2700000",
+		"nginx.org/hsts-include-subdomains": "True",
+	}
+	removedAnnotations := filterMinionAnnotations(minionAnnotations)
+
+	expectedfilteredMinionAnnotations := map[string]string{
+		"nginx.org/rewrites":     "serviceName=service1 rewrite=rewrite1",
+		"nginx.org/ssl-services": "service1",
+	}
+	expectedRemovedAnnotations := []string{
+		"nginx.org/hsts",
+		"nginx.org/hsts-max-age",
+		"nginx.org/hsts-include-subdomains",
+	}
+
+	sort.Strings(removedAnnotations)
+	sort.Strings(expectedRemovedAnnotations)
+
+	if !reflect.DeepEqual(expectedfilteredMinionAnnotations, minionAnnotations) {
+		t.Errorf("filterMinionAnnotations returned %v, but expected %v", minionAnnotations, expectedfilteredMinionAnnotations)
+	}
+	if !reflect.DeepEqual(expectedRemovedAnnotations, removedAnnotations) {
+		t.Errorf("filterMinionAnnotations returned %v, but expected %v", removedAnnotations, expectedRemovedAnnotations)
+	}
+}
+
+func TestMergeMasterAnnotationsIntoMinion(t *testing.T) {
+	masterAnnotations := map[string]string{
+		"nginx.org/proxy-buffering":       "True",
+		"nginx.org/proxy-buffers":         "2",
+		"nginx.org/proxy-buffer-size":     "8k",
+		"nginx.org/hsts":                  "True",
+		"nginx.org/hsts-max-age":          "2700000",
+		"nginx.org/proxy-connect-timeout": "50s",
+		"nginx.com/jwt-token":             "$cookie_auth_token",
+	}
+	minionAnnotations := map[string]string{
+		"nginx.org/client-max-body-size":  "2m",
+		"nginx.org/proxy-connect-timeout": "20s",
+	}
+	mergeMasterAnnotationsIntoMinion(minionAnnotations, masterAnnotations)
+
+	expectedMergedAnnotations := map[string]string{
+		"nginx.org/proxy-buffering":       "True",
+		"nginx.org/proxy-buffers":         "2",
+		"nginx.org/proxy-buffer-size":     "8k",
+		"nginx.org/client-max-body-size":  "2m",
+		"nginx.org/proxy-connect-timeout": "20s",
+	}
+	if !reflect.DeepEqual(expectedMergedAnnotations, minionAnnotations) {
+		t.Errorf("mergeMasterAnnotationsIntoMinion returned %v, but expected %v", minionAnnotations, expectedMergedAnnotations)
+	}
+}
+
+func TestParseSlowStart(t *testing.T) {
+	var testsWithValidInput = []string{"1", "1m10s", "11 11", "5m 30s", "1s", "100m", "5w", "15m", "11M", "3h", "100y", "600"}
+	var invalidInput = []string{"ss", "rM", "m0m", "s1s", "-5s", "", "1L"}
+	for _, test := range testsWithValidInput {
+		result, err := parseSlowStart(test)
+		if err != nil {
+			t.Errorf("TestParseSlowStart(%q) returned an error for valid input", test)
+		}
+		if test != result {
+			t.Errorf("TestParseSlowStart(%q) returned %q expected %q", test, result, test)
+		}
+	}
+	for _, test := range invalidInput {
+		result, err := parseSlowStart(test)
+		if err == nil {
+			t.Errorf("TestParseSlowStart(%q) didn't return error. Returned: %q", test, result)
+		}
+	}
+}
