+++ internal/configs/configmaps.go
@@ -4,123 +4,25 @@ import (
 	"strings"
 
 	"github.com/golang/glog"
-	api_v1 "k8s.io/api/core/v1"
+	"github.com/nginxinc/kubernetes-ingress/internal/configs/version1"
+	"k8s.io/api/core/v1"
 )
 
-// Config holds NGINX configuration parameters
-type Config struct {
-	LocationSnippets              []string
-	ServerSnippets                []string
-	ServerTokens                  string
-	ProxyConnectTimeout           string
-	ProxyReadTimeout              string
-	ClientMaxBodySize             string
-	HTTP2                         bool
-	RedirectToHTTPS               bool
-	SSLRedirect                   bool
-	MainMainSnippets              []string
-	MainHTTPSnippets              []string
-	MainStreamSnippets            []string
-	MainServerNamesHashBucketSize string
-	MainServerNamesHashMaxSize    string
-	MainAccessLogOff              bool
-	MainLogFormat                 string
-	MainErrorLogLevel             string
-	MainStreamLogFormat           string
-	ProxyBuffering                bool
-	ProxyBuffers                  string
-	ProxyBufferSize               string
-	ProxyMaxTempFileSize          string
-	ProxyProtocol                 bool
-	ProxyHideHeaders              []string
-	ProxyPassHeaders              []string
-	HSTS                          bool
-	HSTSBehindProxy               bool
-	HSTSMaxAge                    int64
-	HSTSIncludeSubdomains         bool
-	LBMethod                      string
-	MainWorkerProcesses           string
-	MainWorkerCPUAffinity         string
-	MainWorkerShutdownTimeout     string
-	MainWorkerConnections         string
-	MainWorkerRlimitNofile        string
-	Keepalive                     int64
-	MaxFails                      int
-	FailTimeout                   string
-	HealthCheckEnabled            bool
-	HealthCheckMandatory          bool
-	HealthCheckMandatoryQueue     int64
-	SlowStart                     string
-	ResolverAddresses             []string
-	ResolverIPV6                  bool
-	ResolverValid                 string
-	ResolverTimeout               string
-	MainKeepaliveTimeout          string
-	MainKeepaliveRequests         int64
+// ParseConfigMap parses ConfigMap into ConfigParams.
+func ParseConfigMap(cfgm *v1.ConfigMap, nginxPlus bool) *ConfigParams {
+	cfgParams := NewDefaultConfigParams()
 
-	// http://nginx.org/en/docs/http/ngx_http_realip_module.html
-	RealIPHeader    string
-	SetRealIPFrom   []string
-	RealIPRecursive bool
-
-	// http://nginx.org/en/docs/http/ngx_http_ssl_module.html
-	MainServerSSLProtocols           string
-	MainServerSSLPreferServerCiphers bool
-	MainServerSSLCiphers             string
-	MainServerSSLDHParam             string
-	MainServerSSLDHParamFileContent  *string
-
-	MainTemplate    *string
-	IngressTemplate *string
-
-	JWTRealm    string
-	JWTKey      string
-	JWTToken    string
-	JWTLoginURL string
-
-	Ports    []int
-	SSLPorts []int
-}
-
-// NewDefaultConfig creates a Config with default values
-func NewDefaultConfig() *Config {
-	return &Config{
-		ServerTokens:               "on",
-		ProxyConnectTimeout:        "60s",
-		ProxyReadTimeout:           "60s",
-		ClientMaxBodySize:          "1m",
-		SSLRedirect:                true,
-		MainServerNamesHashMaxSize: "512",
-		ProxyBuffering:             true,
-		MainWorkerProcesses:        "auto",
-		MainWorkerConnections:      "1024",
-		HSTSMaxAge:                 2592000,
-		Ports:                      []int{80},
-		SSLPorts:                   []int{443},
-		MaxFails:                   1,
-		FailTimeout:                "10s",
-		LBMethod:                   "random two least_conn",
-		MainErrorLogLevel:          "notice",
-		ResolverIPV6:               true,
-		MainKeepaliveTimeout:       "65s",
-		MainKeepaliveRequests:      100,
-	}
-}
-
-// ParseConfigMap Parse ConfigMap to Config
-func ParseConfigMap(cfgm *api_v1.ConfigMap, nginxPlus bool) *Config {
-	cfg := NewDefaultConfig()
 	if serverTokens, exists, err := GetMapKeyAsBool(cfgm.Data, "server-tokens", cfgm); exists {
 		if err != nil {
 			if nginxPlus {
-				cfg.ServerTokens = cfgm.Data["server-tokens"]
+				cfgParams.ServerTokens = cfgm.Data["server-tokens"]
 			} else {
 				glog.Error(err)
 			}
 		} else {
-			cfg.ServerTokens = "off"
+			cfgParams.ServerTokens = "off"
 			if serverTokens {
-				cfg.ServerTokens = "on"
+				cfgParams.ServerTokens = "on"
 			}
 		}
 	}
@@ -130,69 +32,77 @@ func ParseConfigMap(cfgm *api_v1.ConfigMap, nginxPlus bool) *Config {
 			if parsedMethod, err := ParseLBMethodForPlus(lbMethod); err != nil {
 				glog.Errorf("Configmap %s/%s: Invalid value for the lb-method key: got %q: %v", cfgm.GetNamespace(), cfgm.GetName(), lbMethod, err)
 			} else {
-				cfg.LBMethod = parsedMethod
+				cfgParams.LBMethod = parsedMethod
 			}
 		} else {
 			if parsedMethod, err := ParseLBMethod(lbMethod); err != nil {
 				glog.Errorf("Configmap %s/%s: Invalid value for the lb-method key: got %q: %v", cfgm.GetNamespace(), cfgm.GetName(), lbMethod, err)
 			} else {
-				cfg.LBMethod = parsedMethod
+				cfgParams.LBMethod = parsedMethod
 			}
 		}
 	}
 
 	if proxyConnectTimeout, exists := cfgm.Data["proxy-connect-timeout"]; exists {
-		cfg.ProxyConnectTimeout = proxyConnectTimeout
+		cfgParams.ProxyConnectTimeout = proxyConnectTimeout
 	}
+
 	if proxyReadTimeout, exists := cfgm.Data["proxy-read-timeout"]; exists {
-		cfg.ProxyReadTimeout = proxyReadTimeout
+		cfgParams.ProxyReadTimeout = proxyReadTimeout
 	}
+
 	if proxyHideHeaders, exists, err := GetMapKeyAsStringSlice(cfgm.Data, "proxy-hide-headers", cfgm, ","); exists {
 		if err != nil {
 			glog.Error(err)
 		} else {
-			cfg.ProxyHideHeaders = proxyHideHeaders
+			cfgParams.ProxyHideHeaders = proxyHideHeaders
 		}
 	}
+
 	if proxyPassHeaders, exists, err := GetMapKeyAsStringSlice(cfgm.Data, "proxy-pass-headers", cfgm, ","); exists {
 		if err != nil {
 			glog.Error(err)
 		} else {
-			cfg.ProxyPassHeaders = proxyPassHeaders
+			cfgParams.ProxyPassHeaders = proxyPassHeaders
 		}
 	}
+
 	if clientMaxBodySize, exists := cfgm.Data["client-max-body-size"]; exists {
-		cfg.ClientMaxBodySize = clientMaxBodySize
+		cfgParams.ClientMaxBodySize = clientMaxBodySize
 	}
+
 	if serverNamesHashBucketSize, exists := cfgm.Data["server-names-hash-bucket-size"]; exists {
-		cfg.MainServerNamesHashBucketSize = serverNamesHashBucketSize
+		cfgParams.MainServerNamesHashBucketSize = serverNamesHashBucketSize
 	}
+
 	if serverNamesHashMaxSize, exists := cfgm.Data["server-names-hash-max-size"]; exists {
-		cfg.MainServerNamesHashMaxSize = serverNamesHashMaxSize
+		cfgParams.MainServerNamesHashMaxSize = serverNamesHashMaxSize
 	}
+
 	if HTTP2, exists, err := GetMapKeyAsBool(cfgm.Data, "http2", cfgm); exists {
 		if err != nil {
 			glog.Error(err)
 		} else {
-			cfg.HTTP2 = HTTP2
+			cfgParams.HTTP2 = HTTP2
 		}
 	}
+
 	if redirectToHTTPS, exists, err := GetMapKeyAsBool(cfgm.Data, "redirect-to-https", cfgm); exists {
 		if err != nil {
 			glog.Error(err)
 		} else {
-			cfg.RedirectToHTTPS = redirectToHTTPS
+			cfgParams.RedirectToHTTPS = redirectToHTTPS
 		}
 	}
+
 	if sslRedirect, exists, err := GetMapKeyAsBool(cfgm.Data, "ssl-redirect", cfgm); exists {
 		if err != nil {
 			glog.Error(err)
 		} else {
-			cfg.SSLRedirect = sslRedirect
+			cfgParams.SSLRedirect = sslRedirect
 		}
 	}
 
-	// HSTS block
 	if hsts, exists, err := GetMapKeyAsBool(cfgm.Data, "hsts", cfgm); exists {
 		if err != nil {
 			glog.Error(err)
@@ -218,15 +128,15 @@ func ParseConfigMap(cfgm *api_v1.ConfigMap, nginxPlus bool) *Config {
 			if parsingErrors {
 				glog.Errorf("Configmap %s/%s: There are configuration issues with hsts annotations, skipping options for all hsts settings", cfgm.GetNamespace(), cfgm.GetName())
 			} else {
-				cfg.HSTS = hsts
+				cfgParams.HSTS = hsts
 				if existsMA {
-					cfg.HSTSMaxAge = hstsMaxAge
+					cfgParams.HSTSMaxAge = hstsMaxAge
 				}
 				if existsIS {
-					cfg.HSTSIncludeSubdomains = hstsIncludeSubdomains
+					cfgParams.HSTSIncludeSubdomains = hstsIncludeSubdomains
 				}
 				if existsBP {
-					cfg.HSTSBehindProxy = hstsBehindProxy
+					cfgParams.HSTSBehindProxy = hstsBehindProxy
 				}
 			}
 		}
@@ -236,155 +146,180 @@ func ParseConfigMap(cfgm *api_v1.ConfigMap, nginxPlus bool) *Config {
 		if err != nil {
 			glog.Error(err)
 		} else {
-			cfg.ProxyProtocol = proxyProtocol
+			cfgParams.ProxyProtocol = proxyProtocol
 		}
 	}
 
-	// ngx_http_realip_module
 	if realIPHeader, exists := cfgm.Data["real-ip-header"]; exists {
-		cfg.RealIPHeader = realIPHeader
+		cfgParams.RealIPHeader = realIPHeader
 	}
+
 	if setRealIPFrom, exists, err := GetMapKeyAsStringSlice(cfgm.Data, "set-real-ip-from", cfgm, ","); exists {
 		if err != nil {
 			glog.Error(err)
 		} else {
-			cfg.SetRealIPFrom = setRealIPFrom
+			cfgParams.SetRealIPFrom = setRealIPFrom
 		}
 	}
+
 	if realIPRecursive, exists, err := GetMapKeyAsBool(cfgm.Data, "real-ip-recursive", cfgm); exists {
 		if err != nil {
 			glog.Error(err)
 		} else {
-			cfg.RealIPRecursive = realIPRecursive
+			cfgParams.RealIPRecursive = realIPRecursive
 		}
 	}
 
-	// SSL block
 	if sslProtocols, exists := cfgm.Data["ssl-protocols"]; exists {
-		cfg.MainServerSSLProtocols = sslProtocols
+		cfgParams.MainServerSSLProtocols = sslProtocols
 	}
+
 	if sslPreferServerCiphers, exists, err := GetMapKeyAsBool(cfgm.Data, "ssl-prefer-server-ciphers", cfgm); exists {
 		if err != nil {
 			glog.Error(err)
 		} else {
-			cfg.MainServerSSLPreferServerCiphers = sslPreferServerCiphers
+			cfgParams.MainServerSSLPreferServerCiphers = sslPreferServerCiphers
 		}
 	}
+
 	if sslCiphers, exists := cfgm.Data["ssl-ciphers"]; exists {
-		cfg.MainServerSSLCiphers = strings.Trim(sslCiphers, "\n")
+		cfgParams.MainServerSSLCiphers = strings.Trim(sslCiphers, "\n")
 	}
+
 	if sslDHParamFile, exists := cfgm.Data["ssl-dhparam-file"]; exists {
 		sslDHParamFile = strings.Trim(sslDHParamFile, "\n")
-		cfg.MainServerSSLDHParamFileContent = &sslDHParamFile
+		cfgParams.MainServerSSLDHParamFileContent = &sslDHParamFile
 	}
+
 	if errorLogLevel, exists := cfgm.Data["error-log-level"]; exists {
-		cfg.MainErrorLogLevel = errorLogLevel
+		cfgParams.MainErrorLogLevel = errorLogLevel
 	}
+
 	if accessLogOff, exists, err := GetMapKeyAsBool(cfgm.Data, "access-log-off", cfgm); exists {
 		if err != nil {
 			glog.Error(err)
 		} else {
-			cfg.MainAccessLogOff = accessLogOff
+			cfgParams.MainAccessLogOff = accessLogOff
 		}
 	}
+
 	if logFormat, exists := cfgm.Data["log-format"]; exists {
-		cfg.MainLogFormat = logFormat
+		cfgParams.MainLogFormat = logFormat
 	}
+
 	if streamLogFormat, exists := cfgm.Data["stream-log-format"]; exists {
-		cfg.MainStreamLogFormat = streamLogFormat
+		cfgParams.MainStreamLogFormat = streamLogFormat
 	}
+
 	if proxyBuffering, exists, err := GetMapKeyAsBool(cfgm.Data, "proxy-buffering", cfgm); exists {
 		if err != nil {
 			glog.Error(err)
 		} else {
-			cfg.ProxyBuffering = proxyBuffering
+			cfgParams.ProxyBuffering = proxyBuffering
 		}
 	}
+
 	if proxyBuffers, exists := cfgm.Data["proxy-buffers"]; exists {
-		cfg.ProxyBuffers = proxyBuffers
+		cfgParams.ProxyBuffers = proxyBuffers
 	}
+
 	if proxyBufferSize, exists := cfgm.Data["proxy-buffer-size"]; exists {
-		cfg.ProxyBufferSize = proxyBufferSize
+		cfgParams.ProxyBufferSize = proxyBufferSize
 	}
+
 	if proxyMaxTempFileSize, exists := cfgm.Data["proxy-max-temp-file-size"]; exists {
-		cfg.ProxyMaxTempFileSize = proxyMaxTempFileSize
+		cfgParams.ProxyMaxTempFileSize = proxyMaxTempFileSize
 	}
 
 	if mainMainSnippets, exists, err := GetMapKeyAsStringSlice(cfgm.Data, "main-snippets", cfgm, "\n"); exists {
 		if err != nil {
 			glog.Error(err)
 		} else {
-			cfg.MainMainSnippets = mainMainSnippets
+			cfgParams.MainMainSnippets = mainMainSnippets
 		}
 	}
+
 	if mainHTTPSnippets, exists, err := GetMapKeyAsStringSlice(cfgm.Data, "http-snippets", cfgm, "\n"); exists {
 		if err != nil {
 			glog.Error(err)
 		} else {
-			cfg.MainHTTPSnippets = mainHTTPSnippets
+			cfgParams.MainHTTPSnippets = mainHTTPSnippets
 		}
 	}
+
 	if locationSnippets, exists, err := GetMapKeyAsStringSlice(cfgm.Data, "location-snippets", cfgm, "\n"); exists {
 		if err != nil {
 			glog.Error(err)
 		} else {
-			cfg.LocationSnippets = locationSnippets
+			cfgParams.LocationSnippets = locationSnippets
 		}
 	}
+
 	if serverSnippets, exists, err := GetMapKeyAsStringSlice(cfgm.Data, "server-snippets", cfgm, "\n"); exists {
 		if err != nil {
 			glog.Error(err)
 		} else {
-			cfg.ServerSnippets = serverSnippets
+			cfgParams.ServerSnippets = serverSnippets
 		}
 	}
+
 	if _, exists, err := GetMapKeyAsInt(cfgm.Data, "worker-processes", cfgm); exists {
 		if err != nil && cfgm.Data["worker-processes"] != "auto" {
 			glog.Errorf("Configmap %s/%s: Invalid value for worker-processes key: must be an integer or the string 'auto', got %q", cfgm.GetNamespace(), cfgm.GetName(), cfgm.Data["worker-processes"])
 		} else {
-			cfg.MainWorkerProcesses = cfgm.Data["worker-processes"]
+			cfgParams.MainWorkerProcesses = cfgm.Data["worker-processes"]
 		}
 	}
+
 	if workerCPUAffinity, exists := cfgm.Data["worker-cpu-affinity"]; exists {
-		cfg.MainWorkerCPUAffinity = workerCPUAffinity
+		cfgParams.MainWorkerCPUAffinity = workerCPUAffinity
 	}
+
 	if workerShutdownTimeout, exists := cfgm.Data["worker-shutdown-timeout"]; exists {
-		cfg.MainWorkerShutdownTimeout = workerShutdownTimeout
+		cfgParams.MainWorkerShutdownTimeout = workerShutdownTimeout
 	}
+
 	if workerConnections, exists := cfgm.Data["worker-connections"]; exists {
-		cfg.MainWorkerConnections = workerConnections
+		cfgParams.MainWorkerConnections = workerConnections
 	}
+
 	if workerRlimitNofile, exists := cfgm.Data["worker-rlimit-nofile"]; exists {
-		cfg.MainWorkerRlimitNofile = workerRlimitNofile
+		cfgParams.MainWorkerRlimitNofile = workerRlimitNofile
 	}
+
 	if keepalive, exists, err := GetMapKeyAsInt64(cfgm.Data, "keepalive", cfgm); exists {
 		if err != nil {
 			glog.Error(err)
 		} else {
-			cfg.Keepalive = keepalive
+			cfgParams.Keepalive = keepalive
 		}
 	}
+
 	if maxFails, exists, err := GetMapKeyAsInt(cfgm.Data, "max-fails", cfgm); exists {
 		if err != nil {
 			glog.Error(err)
 		} else {
-			cfg.MaxFails = maxFails
+			cfgParams.MaxFails = maxFails
 		}
 	}
+
 	if failTimeout, exists := cfgm.Data["fail-timeout"]; exists {
-		cfg.FailTimeout = failTimeout
+		cfgParams.FailTimeout = failTimeout
 	}
+
 	if mainTemplate, exists := cfgm.Data["main-template"]; exists {
-		cfg.MainTemplate = &mainTemplate
+		cfgParams.MainTemplate = &mainTemplate
 	}
+
 	if ingressTemplate, exists := cfgm.Data["ingress-template"]; exists {
-		cfg.IngressTemplate = &ingressTemplate
+		cfgParams.IngressTemplate = &ingressTemplate
 	}
+
 	if mainStreamSnippets, exists, err := GetMapKeyAsStringSlice(cfgm.Data, "stream-snippets", cfgm, "\n"); exists {
 		if err != nil {
 			glog.Error(err)
 		} else {
-			cfg.MainStreamSnippets = mainStreamSnippets
+			cfgParams.MainStreamSnippets = mainStreamSnippets
 		}
 	}
 
@@ -393,7 +328,7 @@ func ParseConfigMap(cfgm *api_v1.ConfigMap, nginxPlus bool) *Config {
 			glog.Error(err)
 		} else {
 			if nginxPlus {
-				cfg.ResolverAddresses = resolverAddresses
+				cfgParams.ResolverAddresses = resolverAddresses
 			} else {
 				glog.Warning("ConfigMap key 'resolver-addresses' requires NGINX Plus")
 			}
@@ -405,7 +340,7 @@ func ParseConfigMap(cfgm *api_v1.ConfigMap, nginxPlus bool) *Config {
 			glog.Error(err)
 		} else {
 			if nginxPlus {
-				cfg.ResolverIPV6 = resolverIpv6
+				cfgParams.ResolverIPV6 = resolverIpv6
 			} else {
 				glog.Warning("ConfigMap key 'resolver-ipv6' requires NGINX Plus")
 			}
@@ -414,7 +349,7 @@ func ParseConfigMap(cfgm *api_v1.ConfigMap, nginxPlus bool) *Config {
 
 	if resolverValid, exists := cfgm.Data["resolver-valid"]; exists {
 		if nginxPlus {
-			cfg.ResolverValid = resolverValid
+			cfgParams.ResolverValid = resolverValid
 		} else {
 			glog.Warning("ConfigMap key 'resolver-valid' requires NGINX Plus")
 		}
@@ -422,22 +357,62 @@ func ParseConfigMap(cfgm *api_v1.ConfigMap, nginxPlus bool) *Config {
 
 	if resolverTimeout, exists := cfgm.Data["resolver-timeout"]; exists {
 		if nginxPlus {
-			cfg.ResolverTimeout = resolverTimeout
+			cfgParams.ResolverTimeout = resolverTimeout
 		} else {
 			glog.Warning("ConfigMap key 'resolver-timeout' requires NGINX Plus")
 		}
 	}
 
 	if keepaliveTimeout, exists := cfgm.Data["keepalive-timeout"]; exists {
-		cfg.MainKeepaliveTimeout = keepaliveTimeout
+		cfgParams.MainKeepaliveTimeout = keepaliveTimeout
 	}
+
 	if keepaliveRequests, exists, err := GetMapKeyAsInt64(cfgm.Data, "keepalive-requests", cfgm); exists {
 		if err != nil {
 			glog.Error(err)
 		} else {
-			cfg.MainKeepaliveRequests = keepaliveRequests
+			cfgParams.MainKeepaliveRequests = keepaliveRequests
 		}
 	}
 
-	return cfg
+	return cfgParams
+}
+
+// GenerateNginxMainConfig generates MainConfig.
+func GenerateNginxMainConfig(staticCfgParams *StaticConfigParams, config *ConfigParams) *version1.MainConfig {
+	nginxCfg := &version1.MainConfig{
+		HealthStatus:                   staticCfgParams.HealthStatus,
+		NginxStatus:                    staticCfgParams.NginxStatus,
+		NginxStatusAllowCIDRs:          staticCfgParams.NginxStatusAllowCIDRs,
+		NginxStatusPort:                staticCfgParams.NginxStatusPort,
+		StubStatusOverUnixSocketForOSS: staticCfgParams.StubStatusOverUnixSocketForOSS,
+		MainSnippets:                   config.MainMainSnippets,
+		HTTPSnippets:                   config.MainHTTPSnippets,
+		StreamSnippets:                 config.MainStreamSnippets,
+		ServerNamesHashBucketSize:      config.MainServerNamesHashBucketSize,
+		ServerNamesHashMaxSize:         config.MainServerNamesHashMaxSize,
+		AccessLogOff:                   config.MainAccessLogOff,
+		LogFormat:                      config.MainLogFormat,
+		ErrorLogLevel:                  config.MainErrorLogLevel,
+		StreamLogFormat:                config.MainStreamLogFormat,
+		SSLProtocols:                   config.MainServerSSLProtocols,
+		SSLCiphers:                     config.MainServerSSLCiphers,
+		SSLDHParam:                     config.MainServerSSLDHParam,
+		SSLPreferServerCiphers:         config.MainServerSSLPreferServerCiphers,
+		HTTP2:                          config.HTTP2,
+		ServerTokens:                   config.ServerTokens,
+		ProxyProtocol:                  config.ProxyProtocol,
+		WorkerProcesses:                config.MainWorkerProcesses,
+		WorkerCPUAffinity:              config.MainWorkerCPUAffinity,
+		WorkerShutdownTimeout:          config.MainWorkerShutdownTimeout,
+		WorkerConnections:              config.MainWorkerConnections,
+		WorkerRlimitNofile:             config.MainWorkerRlimitNofile,
+		ResolverAddresses:              config.ResolverAddresses,
+		ResolverIPV6:                   config.ResolverIPV6,
+		ResolverValid:                  config.ResolverValid,
+		ResolverTimeout:                config.ResolverTimeout,
+		KeepaliveTimeout:               config.MainKeepaliveTimeout,
+		KeepaliveRequests:              config.MainKeepaliveRequests,
+	}
+	return nginxCfg
 }
