+++ internal/configs/parsing_helpers_test.go
@@ -4,12 +4,12 @@ import (
 	"reflect"
 	"testing"
 
-	api_v1 "k8s.io/api/core/v1"
-	extensions "k8s.io/api/extensions/v1beta1"
+	"k8s.io/api/core/v1"
+	"k8s.io/api/extensions/v1beta1"
 	meta_v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
-var configMap = api_v1.ConfigMap{
+var configMap = v1.ConfigMap{
 	ObjectMeta: meta_v1.ObjectMeta{
 		Name:      "test",
 		Namespace: "default",
@@ -19,7 +19,7 @@ var configMap = api_v1.ConfigMap{
 		APIVersion: "v1",
 	},
 }
-var ingress = extensions.Ingress{
+var ingress = v1beta1.Ingress{
 	ObjectMeta: meta_v1.ObjectMeta{
 		Name:      "test",
 		Namespace: "kube-system",
@@ -30,9 +30,6 @@ var ingress = extensions.Ingress{
 	},
 }
 
-//
-// GetMapKeyAsBool
-//
 func TestGetMapKeyAsBool(t *testing.T) {
 	configMap := configMap
 	configMap.Data = map[string]string{
@@ -82,6 +79,7 @@ func TestGetMapKeyAsBoolErrorMessage(t *testing.T) {
 	ingress.Annotations = map[string]string{
 		"key": "other_string",
 	}
+
 	_, _, err = GetMapKeyAsBool(ingress.Annotations, "key", &ingress)
 	if err == nil {
 		t.Error("An error was expected")
@@ -92,9 +90,6 @@ func TestGetMapKeyAsBoolErrorMessage(t *testing.T) {
 	}
 }
 
-//
-// GetMapKeyAsInt
-//
 func TestGetMapKeyAsInt(t *testing.T) {
 	configMap := configMap
 	configMap.Data = map[string]string{
@@ -145,6 +140,7 @@ func TestGetMapKeyAsIntErrorMessage(t *testing.T) {
 	ingress.Annotations = map[string]string{
 		"key": "other_string",
 	}
+
 	_, _, err = GetMapKeyAsInt(ingress.Annotations, "key", &ingress)
 	if err == nil {
 		t.Error("An error was expected")
@@ -155,9 +151,6 @@ func TestGetMapKeyAsIntErrorMessage(t *testing.T) {
 	}
 }
 
-//
-// GetMapKeyAsInt64
-//
 func TestGetMapKeyAsInt64(t *testing.T) {
 	configMap := configMap
 	configMap.Data = map[string]string{
@@ -208,6 +201,7 @@ func TestGetMapKeyAsInt64ErrorMessage(t *testing.T) {
 	ingress.Annotations = map[string]string{
 		"key": "other_string",
 	}
+
 	_, _, err = GetMapKeyAsInt64(ingress.Annotations, "key", &ingress)
 	if err == nil {
 		t.Error("An error was expected")
@@ -218,9 +212,6 @@ func TestGetMapKeyAsInt64ErrorMessage(t *testing.T) {
 	}
 }
 
-//
-// GetMapKeyAsStringSlice
-//
 func TestGetMapKeyAsStringSlice(t *testing.T) {
 	configMap := configMap
 	configMap.Data = map[string]string{
@@ -250,6 +241,7 @@ func TestGetMapKeyAsStringSliceMultilineSnippets(t *testing.T) {
 				rewrite ^ $new_uri permanent;
 			}`,
 	}
+
 	slice, exists, err := GetMapKeyAsStringSlice(configMap.Data, "server-snippets", &configMap, "\n")
 	if err != nil {
 		t.Errorf("Unexpected error: %v", err)
@@ -271,5 +263,102 @@ func TestGetMapKeyAsStringSliceNotFound(t *testing.T) {
 	_, exists, _ := GetMapKeyAsStringSlice(configMap.Data, "key", &configMap, ",")
 	if exists {
 		t.Errorf("The key 'key' must not exist in the configMap")
+	}
+}
+
+func TestParseLBMethod(t *testing.T) {
+	var testsWithValidInput = []struct {
+		input    string
+		expected string
+	}{
+		{"least_conn", "least_conn"},
+		{"round_robin", ""},
+		{"ip_hash", "ip_hash"},
+		{"random", "random"},
+		{"random two", "random two"},
+		{"random two least_conn", "random two least_conn"},
+		{"hash $request_id", "hash $request_id"},
+		{"hash $request_id consistent", "hash $request_id consistent"},
+	}
+
+	var invalidInput = []string{
+		"",
+		"blabla",
+		"least_time header",
+		"hash123",
+		"hash $request_id conwrongspelling",
+		"random one",
+		"random two least_time=header",
+		"random two least_time=last_byte",
+		"random two ip_hash",
+	}
+
+	for _, test := range testsWithValidInput {
+		result, err := ParseLBMethod(test.input)
+		if err != nil {
+			t.Errorf("TestParseLBMethod(%q) returned an error for valid input", test.input)
+		}
+
+		if result != test.expected {
+			t.Errorf("TestParseLBMethod(%q) returned %q expected %q", test.input, result, test.expected)
+		}
+	}
+
+	for _, input := range invalidInput {
+		_, err := ParseLBMethod(input)
+		if err == nil {
+			t.Errorf("TestParseLBMethod(%q) does not return an error for invalid input", input)
+		}
+	}
+}
+
+func TestParseLBMethodForPlus(t *testing.T) {
+	var testsWithValidInput = []struct {
+		input    string
+		expected string
+	}{
+		{"least_conn", "least_conn"},
+		{"round_robin", ""},
+		{"ip_hash", "ip_hash"},
+		{"random", "random"},
+		{"random two", "random two"},
+		{"random two least_conn", "random two least_conn"},
+		{"random two least_time=header", "random two least_time=header"},
+		{"random two least_time=last_byte", "random two least_time=last_byte"},
+		{"hash $request_id", "hash $request_id"},
+		{"least_time header", "least_time header"},
+		{"least_time last_byte", "least_time last_byte"},
+		{"least_time header inflight", "least_time header inflight"},
+		{"least_time last_byte inflight", "least_time last_byte inflight"},
+	}
+
+	var invalidInput = []string{
+		"",
+		"blabla",
+		"hash123",
+		"least_time",
+		"last_byte",
+		"least_time inflight header",
+		"random one",
+		"random two ip_hash",
+		"random two least_time",
+	}
+
+	for _, test := range testsWithValidInput {
+		result, err := ParseLBMethodForPlus(test.input)
+		if err != nil {
+			t.Errorf("TestParseLBMethod(%q) returned an error for valid input", test.input)
+		}
+
+		if result != test.expected {
+			t.Errorf("TestParseLBMethod(%q) returned %q expected %q", test.input, result, test.expected)
+		}
+	}
+
+	for _, input := range invalidInput {
+		_, err := ParseLBMethodForPlus(input)
+		if err == nil {
+			t.Errorf("TestParseLBMethod(%q) does not return an error for invalid input", input)
+		}
 	}
 }
