+++ internal/configs/version1/nginx-plus.ingress.tmpl
