+++ internal/configs/version1/nginx-plus.tmpl
