+++ internal/configs/version1/nginx.ingress.tmpl
