+++ internal/configs/version1/nginx.tmpl
