+++ internal/k8s/secret.go
@@ -1,10 +1,13 @@
-package configs
+package k8s
 
 import (
 	"fmt"
 
-	api_v1 "k8s.io/api/core/v1"
+	"k8s.io/api/core/v1"
 )
+
+// JWTKeyKey is the key of the data field of a Secret where the JWK must be stored.
+const JWTKeyKey = "jwk"
 
 const (
 	// TLS Secret
@@ -14,20 +17,20 @@ const (
 )
 
 // ValidateTLSSecret validates the secret. If it is valid, the function returns nil.
-func ValidateTLSSecret(secret *api_v1.Secret) error {
-	if _, exists := secret.Data[api_v1.TLSCertKey]; !exists {
-		return fmt.Errorf("Secret doesn't have %v", api_v1.TLSCertKey)
+func ValidateTLSSecret(secret *v1.Secret) error {
+	if _, exists := secret.Data[v1.TLSCertKey]; !exists {
+		return fmt.Errorf("Secret doesn't have %v", v1.TLSCertKey)
 	}
 
-	if _, exists := secret.Data[api_v1.TLSPrivateKeyKey]; !exists {
-		return fmt.Errorf("Secret doesn't have %v", api_v1.TLSPrivateKeyKey)
+	if _, exists := secret.Data[v1.TLSPrivateKeyKey]; !exists {
+		return fmt.Errorf("Secret doesn't have %v", v1.TLSPrivateKeyKey)
 	}
 
 	return nil
 }
 
 // ValidateJWKSecret validates the secret. If it is valid, the function returns nil.
-func ValidateJWKSecret(secret *api_v1.Secret) error {
+func ValidateJWKSecret(secret *v1.Secret) error {
 	if _, exists := secret.Data[JWTKeyKey]; !exists {
 		return fmt.Errorf("Secret doesn't have %v", JWTKeyKey)
 	}
@@ -36,7 +39,7 @@ func ValidateJWKSecret(secret *api_v1.Secret) error {
 }
 
 // GetSecretKind returns the kind of the Secret.
-func GetSecretKind(secret *api_v1.Secret) (int, error) {
+func GetSecretKind(secret *v1.Secret) (int, error) {
 	if err := ValidateTLSSecret(secret); err == nil {
 		return TLS, nil
 	}
